backend/requirements.txt
# Core dependencies
python-docx
google-generativeai
fastapi
uvicorn
python-multipart

# PDF processing
PyPDF2
pdf2image
pdf2docx

# OCR (EasyOCR)
easyocr

# Image processing
Pillow

# Utilities
python-dotenv
